+++ src/index.js
@@ -1,19 +1,16 @@
 // Import vue component
 import component from './vue-simple-context-menu.vue';
 
-// install function executed by app.use()
-export function install (app) {
+export function install(app) {
   if (install.installed) return;
+
   install.installed = true;
   app.component('VueSimpleContextMenu', component);
 }
 
-// Create module definition for Vue.use()
-const plugin = {
-  install,
-};
+const plugin = { install };
 
-// To auto-install when vue is found
+// To auto-install when Vue is found
 let GlobalVue = null;
 if (typeof window !== 'undefined') {
   GlobalVue = window.Vue;
@@ -24,7 +21,7 @@ if (GlobalVue) {
   GlobalVue.use(plugin);
 }
 
-// To allow use as module (npm/webpack/etc.) export component
+// To allow usage as a module (npm/webpack/etc.) export component
 export default component;
 
 // It's possible to expose named exports when writing components that can