+++ src/vue-simple-context-menu.vue
@@ -1,53 +1,43 @@
 <template>
   <div>
-    <ul
-      :id="elementId"
-      class="vue-simple-context-menu"
-      v-click-outside="onClickOutside"
-    >
+    <ul :id="elementId" class="vue-simple-context-menu" v-click-outside="onClickOutside">
       <li
         v-for="(option, index) in options"
         :key="index"
         @click.stop="optionClicked(option)"
         class="vue-simple-context-menu__item"
-        :class="[
-          option.class,
-          option.type === 'divider' ? 'vue-simple-context-menu__divider' : ''
-        ]"
+        :class="[option.class, option.type === 'divider' ? 'vue-simple-context-menu__divider' : '']"
       >
-        <span v-html="option.name"></span>
+        <span v-html="option.name" />
       </li>
     </ul>
   </div>
 </template>
 
 <script>
-import vClickOutside from "click-outside-vue3";
+import vClickOutside from 'click-outside-vue3';
 
 export default {
-  name: "VueSimpleContextMenu",
+  name: 'VueSimpleContextMenu',
   props: {
     elementId: {
       type: String,
-      required: true
+      required: true,
     },
     options: {
       type: Array,
-      required: true
-    }
+      required: true,
+    },
   },
-  emits: [
-    'menu-closed',
-    'option-clicked'
-  ],
+  emits: ['menu-closed', 'option-clicked'],
   directives: {
-    'click-outside': vClickOutside.directive
+    'click-outside': vClickOutside.directive,
   },
   data() {
     return {
       item: null,
+      menuHeight: null,
       menuWidth: null,
-      menuHeight: null
     };
   },
   methods: {
@@ -60,32 +50,32 @@ export default {
       }
 
       if (!this.menuWidth || !this.menuHeight) {
-        menu.style.visibility = "hidden";
-        menu.style.display = "block";
+        menu.style.visibility = 'hidden';
+        menu.style.display = 'block';
         this.menuWidth = menu.offsetWidth;
         this.menuHeight = menu.offsetHeight;
-        menu.removeAttribute("style");
+        menu.removeAttribute('style');
       }
 
       if (this.menuWidth + event.pageX >= window.innerWidth) {
-        menu.style.left = event.pageX - this.menuWidth + 2 + "px";
+        menu.style.left = event.pageX - this.menuWidth + 2 + 'px';
       } else {
-        menu.style.left = event.pageX - 2 + "px";
+        menu.style.left = event.pageX - 2 + 'px';
       }
 
       if (this.menuHeight + event.pageY >= window.innerHeight) {
-        menu.style.top = event.pageY - this.menuHeight + 2 + "px";
+        menu.style.top = event.pageY - this.menuHeight + 2 + 'px';
       } else {
-        menu.style.top = event.pageY - 2 + "px";
+        menu.style.top = event.pageY - 2 + 'px';
       }
 
-      menu.classList.add("vue-simple-context-menu--active");
+      menu.classList.add('vue-simple-context-menu--active');
     },
     hideContextMenu() {
       const element = document.getElementById(this.elementId);
       if (element) {
-        element.classList.remove("vue-simple-context-menu--active");
-        this.$emit("menu-closed");
+        element.classList.remove('vue-simple-context-menu--active');
+        this.$emit('menu-closed');
       }
     },
     onClickOutside() {
@@ -93,23 +83,23 @@ export default {
     },
     optionClicked(option) {
       this.hideContextMenu();
-      this.$emit("option-clicked", {
+      this.$emit('option-clicked', {
         item: this.item,
-        option: option
+        option: option,
       });
     },
     onEscKeyRelease(event) {
       if (event.keyCode === 27) {
         this.hideContextMenu();
       }
-    }
+    },
   },
   mounted() {
-    document.body.addEventListener("keyup", this.onEscKeyRelease);
+    document.body.addEventListener('keyup', this.onEscKeyRelease);
   },
   beforeUnmount() {
-    document.removeEventListener("keyup", this.onEscKeyRelease);
-  }
+    document.removeEventListener('keyup', this.onEscKeyRelease);
+  },
 };
 </script>
 
@@ -121,32 +111,31 @@ $white: #fff;
 $black: #333;
 
 .vue-simple-context-menu {
-  top: 0;
-  left: 0;
-  margin: 0;
-  padding: 0;
-  display: none;
-  list-style: none;
-  position: absolute;
-  z-index: 1000000;
   background-color: $light-grey;
   border-bottom-width: 0px;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  box-shadow: 0 3px 6px 0 rgba($black, 0.2);
   border-radius: 4px;
+  box-shadow: 0 3px 6px 0 rgba($black, 0.2);
+  display: none;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
+    'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  left: 0;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  position: absolute;
+  top: 0;
+  z-index: 1000000;
 
   &--active {
     display: block;
   }
 
   &__item {
-    display: flex;
+    align-items: center;
     color: $black;
     cursor: pointer;
+    display: flex;
     padding: 5px 15px;
-    align-items: center;
 
     &:hover {
       background-color: $blue;
@@ -155,16 +144,15 @@ $black: #333;
   }
 
   &__divider {
+    background-clip: content-box;
+    background-color: $grey;
     box-sizing: content-box;
     height: 2px;
-    background-color: $grey;
     padding: 4px 0;
-    background-clip: content-box;
     pointer-events: none;
   }
 
-  // Have to use the element so we can make use of `first-of-type` and
-  // `last-of-type`
+  // Have to use the element so we can make use of `first-of-type` and `last-of-type`
   li {
     &:first-of-type {
       margin-top: 4px;