+++ test/VueSimpleContextMenu.spec.js
@@ -1,5 +1,5 @@
-import { shallowMount } from '@vue/test-utils'
-import VueSimpleContextMenu from '@/vue-simple-context-menu.vue'
+import { shallowMount } from '@vue/test-utils';
+import VueSimpleContextMenu from '@/vue-simple-context-menu.vue';
 
 describe('VueSimpleContextMenu.vue', () => {
   it('Sets props correctly', async () => {
@@ -8,29 +8,29 @@ describe('VueSimpleContextMenu.vue', () => {
       options: [
         {
           name: 'Duplicate',
-          slug: 'duplicate'
+          slug: 'duplicate',
         },
         {
           name: 'Edit',
-          slug: 'edit'
+          slug: 'edit',
         },
         {
           name: 'Delete',
-          slug: 'delete'
-        }
-      ]
-    }
+          slug: 'delete',
+        },
+      ],
+    };
 
     const wrapper = shallowMount(VueSimpleContextMenu, {
       props: {
         elementId: initialPropsData.elementId,
-        options: initialPropsData.options
-      }
-    })
+        options: initialPropsData.options,
+      },
+    });
 
-    expect(wrapper.vm.elementId).toBe(initialPropsData.elementId)
-    expect(wrapper.vm.options).toEqual(initialPropsData.options)
-  })
+    expect(wrapper.vm.elementId).toBe(initialPropsData.elementId);
+    expect(wrapper.vm.options).toEqual(initialPropsData.options);
+  });
 
   it('Shows menu on click', async () => {
     let initialPropsData = {
@@ -38,37 +38,37 @@ describe('VueSimpleContextMenu.vue', () => {
       options: [
         {
           name: 'Duplicate',
-          slug: 'duplicate'
+          slug: 'duplicate',
         },
         {
           name: 'Edit',
-          slug: 'edit'
+          slug: 'edit',
         },
         {
           name: 'Delete',
-          slug: 'delete'
-        }
-      ]
-    }
+          slug: 'delete',
+        },
+      ],
+    };
 
     const wrapper = shallowMount(VueSimpleContextMenu, {
       propsData: {
         elementId: initialPropsData.elementId,
-        options: initialPropsData.options
-      }
-    })
+        options: initialPropsData.options,
+      },
+    });
 
     // Make some test data
-    var testEvent = new Event("click", { "bubbles": true, "cancelable": false });
-    let testItem = { name: 'Jim', job: 'Salesman' }
+    var testEvent = new Event('click', { bubbles: true, cancelable: false });
+    let testItem = { name: 'Jim', job: 'Salesman' };
 
     // Trigger the showing of the menu
-    wrapper.vm.showMenu(testEvent, testItem)
+    wrapper.vm.showMenu(testEvent, testItem);
 
     // Menu show be showing our selected item
-    expect(wrapper.vm.item.name).toBe('Jim')
-    expect(wrapper.vm.item.name).toBe('Jim')
-  })
+    expect(wrapper.vm.item.name).toBe('Jim');
+    expect(wrapper.vm.item.name).toBe('Jim');
+  });
 
   it('Emits event on menu item selection', async () => {
     let initialPropsData = {
@@ -76,42 +76,42 @@ describe('VueSimpleContextMenu.vue', () => {
       options: [
         {
           name: 'Duplicate',
-          slug: 'duplicate'
+          slug: 'duplicate',
         },
         {
           name: 'Edit',
-          slug: 'edit'
+          slug: 'edit',
         },
         {
           name: 'Delete',
-          slug: 'delete'
-        }
-      ]
-    }
+          slug: 'delete',
+        },
+      ],
+    };
 
     const wrapper = shallowMount(VueSimpleContextMenu, {
       propsData: {
         elementId: initialPropsData.elementId,
-        options: initialPropsData.options
-      }
-    })
+        options: initialPropsData.options,
+      },
+    });
 
     // Make some test data
-    var testEvent = new Event("click", { "bubbles": true, "cancelable": false });
-    let testItem = { name: 'Jim', job: 'Salesman' }
+    var testEvent = new Event('click', { bubbles: true, cancelable: false });
+    let testItem = { name: 'Jim', job: 'Salesman' };
 
     // Trigger the showing of the menu
-    wrapper.vm.showMenu(testEvent, testItem)
+    wrapper.vm.showMenu(testEvent, testItem);
 
     // Menu show be showing our selected item
-    expect(wrapper.vm.item.name).toBe('Jim')
-    expect(wrapper.vm.item.job).toBe('Salesman')
+    expect(wrapper.vm.item.name).toBe('Jim');
+    expect(wrapper.vm.item.job).toBe('Salesman');
 
     // Manually click an item on the menu
     // Here we know the options because we set them earlier
-    wrapper.vm.optionClicked(initialPropsData[0])
+    wrapper.vm.optionClicked(initialPropsData[0]);
 
     // Check the event was emitted properly
-    expect(wrapper.emitted('option-clicked')).toBeTruthy()
-  })
-})
+    expect(wrapper.emitted('option-clicked')).toBeTruthy();
+  });
+});