+++ webpack.config.js
@@ -1,16 +1,18 @@
-const path = require('path')
-const webpack = require('webpack')
+// This webpack config is used to serve the example app in ./example
+
+const { VueLoaderPlugin } = require('vue-loader');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const path = require('path');
 const TerserPlugin = require('terser-webpack-plugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin')
-const { VueLoaderPlugin } = require("vue-loader");
-const MiniCssExtractPlugin = require('mini-css-extract-plugin')
+const webpack = require('webpack');
 
 module.exports = {
   entry: './example/main.js',
   output: {
+    filename: 'build.js',
     path: path.resolve(__dirname, './docs'),
-    publicPath: (process.env.NODE_ENV === 'development') ? '/' : '/vue-simple-context-menu/',
-    filename: 'build.js'
+    publicPath: process.env.NODE_ENV === 'development' ? '/' : '/vue-simple-context-menu/',
   },
   module: {
     rules: [
@@ -18,100 +20,100 @@ module.exports = {
         test: /\.css$/,
         use: [
           {
-            loader: MiniCssExtractPlugin.loader
+            loader: MiniCssExtractPlugin.loader,
           },
-          'css-loader'
+          'css-loader',
         ],
       },
       {
         test: /\.scss$/,
         use: [
           {
-            loader: MiniCssExtractPlugin.loader
+            loader: MiniCssExtractPlugin.loader,
           },
           'css-loader',
-          'sass-loader'
+          'sass-loader',
         ],
       },
       {
         test: /\.sass$/,
         use: [
           {
-            loader: MiniCssExtractPlugin.loader
+            loader: MiniCssExtractPlugin.loader,
           },
           'css-loader',
-          'sass-loader?indentedSyntax'
+          'sass-loader?indentedSyntax',
         ],
       },
       {
+        loader: 'vue-loader',
         test: /\.vue$/,
-        loader: 'vue-loader'
       },
       {
-        test: /\.js$/,
+        exclude: /node_modules/,
         loader: 'babel-loader',
-        exclude: /node_modules/
+        test: /\.js$/,
       },
       {
         test: /\.(png|jpg|gif|svg)$/,
         loader: 'file-loader',
         options: {
-          name: '[name].[ext]?[hash]'
-        }
-      }
-    ]
+          name: '[name].[ext]?[hash]',
+        },
+      },
+    ],
   },
   resolve: {
     alias: {
-      'vue$': 'vue/dist/vue.esm-bundler.js'
+      vue$: 'vue/dist/vue.esm-bundler.js',
     },
-    extensions: ['*', '.js', '.vue', '.json']
+    extensions: ['*', '.js', '.vue', '.json'],
   },
   devServer: {
     historyApiFallback: true,
     noInfo: true,
-    overlay: true
+    overlay: true,
   },
   performance: {
-    hints: false
+    hints: false,
   },
   optimization: {
     minimizer: [
       new TerserPlugin({
         parallel: 4,
         terserOptions: {
-          warnings: false,
           compress: {
-            warnings: false
+            warnings: false,
           },
+          warnings: false,
         },
-      })
-    ]
+      }),
+    ],
   },
   devtool: '#eval-source-map',
   plugins: [
     new VueLoaderPlugin(),
     new MiniCssExtractPlugin({
-      filename: '[name].css'
+      filename: '[name].css',
     }),
     new HtmlWebpackPlugin({
       title: 'vue-simple-context-menu',
-      template: './example/index.html'
-    })
-  ]
-}
+      template: './example/index.html',
+    }),
+  ],
+};
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports.devtool = '#source-map'
+  module.exports.devtool = '#source-map';
   // http://vue-loader.vuejs.org/en/workflow/production.html
   module.exports.plugins = (module.exports.plugins || []).concat([
     new webpack.DefinePlugin({
       'process.env': {
-        NODE_ENV: '"production"'
-      }
+        NODE_ENV: '"production"',
+      },
     }),
     new webpack.LoaderOptionsPlugin({
-      minimize: true
-    })
-  ])
+      minimize: true,
+    }),
+  ]);
 }